changed jsx components to mjs and html, working api fetch to add members, redid some styling
