+++ components/classroom/acceptinvite.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <link rel="stylesheet" type="text/css" href="./styles.css" />
+        <title>TPEN Classroom Project Invite</title> 
+    </head>
+    <body>
+        <div class = "header">
+            <h1>TPEN Classroom Interface</h1>
+        </div>
+        <div class="page-container">
+            <h1>You've been invited to a TPEN Classroom Project!</h1>
+            <h2>_user_ has invited you to join _project_ as a _role_</h2>
+            <p>Once you accept, you will be able to contribute based on your assigned role and permissions.</p>
+            <button>Accept Invite</button>
+        </div>
+</body>
