+++ components/classroom/styles.css
@@ -124,6 +124,25 @@ button {
     font-size: 16px;
 }
 
+.modal {
+    display: none;
+    position: fixed;
+    z-index: 1;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    overflow: auto; /* Enable scroll if needed */
+    background-color: rgb(0,0,0,0.4);
+}
+#invite-section-container {
+    background-color: #fefefe;
+    margin: auto;
+    padding: 20px;
+    margin-top: 10px;
+    width: 80%;
+}
+
 /* Hamburger icon */
 .hamburger-menu {
     display: none;
